# 32% Faster Object#try