--- a/activesupport/lib/active_support/core_ext/object/try.rb
+++ b/activesupport/lib/active_support/core_ext/object/try.rb
@@ -5,7 +5,16 @@ require "delegate"
 module ActiveSupport
   module Tryable #:nodoc:
     def try(*a, &b)
-      try!(*a, &b) if a.empty? || respond_to?(a.first)
+      return unless a.empty? || respond_to?(a.first)
+      if a.empty? && block_given?
+        if b.arity == 0
+          instance_eval(&b)
+        else
+          yield self
+        end
+      else
+        public_send(*a, &b)
+      end
     end
 
     def try!(*a, &b)