@matthewd 2018-08-29
I follow the gain from inlining `try!` (though given we generally avoid calling either, especially in hot spots, I'm not sure how valuable it really is).

I don't see why `try!` is changing though. Is that just cosmetic?

> A logical question might be why not get rid of the splat in public_send(*a, &b) for the case where there's only a single symbol and for some reason it's not as big of a savings. No idea why.

I suspect you'll find the saving is in avoiding the splat capture in `try!`, rather than the splat expansion in the call: that's the one that allocates, after all.