@schneems 2018-08-29
> I suspect you'll find the saving is in avoiding the splat capture in try!, rather than the splat expansion in the call: that's the one that allocates, after all.

Ahh, that makes sense. Thanks. My first "DRY" attempt of fixing this was to introduce a private method without the splat, but I like this approach better.

> I don't see why try! is changing though. Is that just cosmetic?

This version was easier for me to see that both versions are identical. I'll change it back to the original logic.