@schneems 2018-08-29
Updated to use original logic.