@sgrif 2018-08-29
There's definitely another 20% win by pulling out the method name into its own argument. Will open a followup PR